# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Tasa-API
  description: "API for synchronize TASA app data."
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /location/create:
    post:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Register a location of a user
      description: Creates a location given coordinates, radius and a name.
      requestBody:
        description: Location data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Location name is empty.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: A location with given name already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/{id}/update/radius/{radius}:
    put:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Update a location
      description: Updates a location given a new radius.
      parameters:
        - name: id
          in: path
          description: Location identifier
          required: true
          schema:
            type: integer
            format: int32
        - name: radius
          in: path
          description: New radius
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Radius is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Location not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/{id}/update/name/{name}:
    put:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Update a location name
      description: Updates a location given a new name.
      parameters:
        - name: id
          in: path
          description: location identifier
          required: true
          schema:
            type: integer
            format: int32
        - name: name
          in: path
          description: location identifier
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Name is blank
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Location not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Location with given name already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Gets a location.
      description: Gets a location by it's identifier.
      parameters:
        - name: id
          in: path
          description: Location's identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Name is blank
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Location not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Gets all locations.
      description: Gets all locations of a user.
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationList'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /location/remove/{id}:
    delete:
      security:
        - bearerAuth: []
      tags:
        - Locations
      summary: Remove a location of a user
      description: Deletes a location given its identifier.
      parameters:
        - name: id
          in: path
          description: Location identifier
          required: true
          schema:
            type: integer
            format: int64
            example: 5208
      responses:
        '200':
          description: Ok
        '400':
          description: Location identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Location not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /rule/location/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Gets a rule location
      description: Gets a rule location by it's identifier.
      parameters:
        - name: id
          in: path
          description: rule identifier
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleLocationOutput'
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      security:
        - bearerAuth: []
      tags:
        - Rules
      summary: Deletes a rule location
      description: Deletes a rule location by it's identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Ok
        '400':
          description: Identifier is negative.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Rule not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LocationInput:
      type: object
      required:
        - latitude
        - longitude
        - name
        - radius
      properties:
        name:
          type: string
          description: Location name.
          example: "Home"
        latitude:
          type: number
          format: double
          description: Latitude of location.
          example: 38.75691186259142
        longitude:
          type: number
          format: double
          description: Langitude of location.
          example: -9.116494150002076
        radius:
          type: number
          format: double
          description: Radius of location.
          example: 30.0
    Location:
      type: object
      required:
        - id
        - latitude
        - longitude
        - name
        - radius
      properties:
        id:
          type: integer
          description: Location dentifier
          example: 10
        name:
          type: string
          description: Location name.
          example: "Home"
        latitude:
          type: number
          format: double
          description: Latitude of location.
          example: 38.75691186259142
        longitude:
          type: number
          format: double
          description: Langitude of location.
          example: -9.116494150002076
        radius:
          type: number
          format: double
          description: Radius of location.
          example: 30.0
    LocationList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/Location'
    RuleLocationOutput:
      required:
        - id
        - endTime
        - locationId
        - startTime
      type: object
      properties:
        id:
          type: integer
          description: Rule dentifier
          example: 10
        startTime:
          type: string
          format: date-time
          description: Start time of rule
          example: "2024-10-01T12:00:00Z"
        endTime:
          type: string
          format: date-time
          description: End time of rule
          example: "2024-10-01T12:00:00Z"
        locationId:
          type: integer
          format: int32
          description: Location identifier.
          example: 10
...
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string